Extract convertMuiFiltersToApi to its own file
